qt6-migration: Add CircularGauge custom QML type
Presently, multiple QML files in this project use CircularGauge and
CircularGaugeStyle, which are QML types defined in Qt-5 modules. These
modules have been removed from Qt-6. Qt-6 does not have a
straightforward alternative to these widgets.

As a result, implement a custom CircularGauge type, and in it integrate
the functionality of Qt-5's CircularGaugeStyle that we need for this
project.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>