qt6-migration: live_camera: Migrate to GStreamer API and appsink
In Qt5, it was possible to rely on MediaPlayer.{stop(), play(),
source()} to run and stop the pipeline, but this feature is now gone.

Further, there also used to be "qtvideosinK" as a sink for GStreamer
pipelines, but this has been deprecated too.

As a result, use appsink as the sink in GStreamer pipeline, and use
QVideoSink for connecting this to the application. Further use GStreamer
APIs for running and stopping pipelines.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>